Add new features and updates: include image, update application properties, add enums and DTOs, implement repositories, and enhance Warehouse model

+++ src/main/java/kz/findmyname284/springbootproject/dto/ProductItemDTO.java
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.NotBlank;
 
 public record ProductItemDTO(
         @NotBlank Long id,
-        @NotBlank Integer quantity) {
+        @NotBlank Long quantity) {
 }
